create patient filters helper
